requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL support
PyMySQL==1.1.0  # MySQL support

# Trading APIs
kiteconnect==4.1.0
yfinance==0.2.28
alpha-vantage==2.3.1  # Additional market data provider

# Telegram Bot
python-telegram-bot==20.7

# Data Analysis and Scientific Computing
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Technical Analysis
ta-lib==0.4.28
talib==0.4.28

# Utilities
python-dotenv==1.0.0
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
schedule==1.2.0
apscheduler==3.10.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Logging and Monitoring
structlog==23.2.0
loguru==0.7.2

# Email Support
smtplib-ssl==1.0.1
email-validator==2.1.0

# Performance and Caching
redis==5.0.1
cachetools==5.3.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.6.0

# Date and Time
python-dateutil==2.8.2
pytz==2023.3
arrow==1.3.0

# Configuration
click==8.1.7
typer==0.9.0  # CLI interface

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Monitoring and Metrics
prometheus-client==0.19.0
psutil==5.9.6

# JSON and Data Serialization
orjson==3.9.10
msgpack==1.0.7

# Environment and System
python-multipart==0.0.6
python-magic==0.4.27

# Additional Database Drivers
aiosqlite==0.19.0  # Async SQLite
asyncpg==0.29.0  # Async PostgreSQL

# Web Scraping (if needed for earnings data)
beautifulsoup4==4.12.2
selenium==4.15.2

# File Handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Timezone Support
zoneinfo==0.2.1; python_version < "3.9"

# Rate Limiting
slowapi==0.1.9
limits==3.6.0

# Validation and Parsing
marshmallow==3.20.1
jsonschema==4.20.0

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Backup and Archiving
boto3==1.34.0  # AWS S3 for backups
google-cloud-storage==2.10.0  # Google Cloud Storage

# Testing and Mocking
factory-boy==3.3.0
faker==20.1.0

# Documentation (optional)
mkdocs==1.5.3
mkdocs-material==9.4.14